update with new tsconfig for documentation and exclude the scripts
