--- documentation/scripts/redirect-index.html
@@ -1,11 +1,0 @@
-<!-- Used to redirect from the base bath ('/packs/') to the latest folder -->
-<!DOCTYPE html>
-<html lang="en" class="no-js">
-  <head>
-    <meta http-equiv="refresh" content="0; url=./latest/" />
-  </head>
-  <!-- We add a body just in case some browsers fail to redirect -->
-  <body>
-    You should be redirected to the docs soon. If not, <a href="./latest/">click here</a>.
-  </body>
-</html>
