+++ tsconfig.scripts.json
@@ -1,3 +1,6 @@
+/*
+ * This is the tsconfig for our documentation scripts.
+ */
 {
   "compilerOptions": {
     /* Basic Options */
@@ -11,13 +14,13 @@
     // "allowJs": true,                          /* Allow javascript files to be compiled. */
     // "checkJs": true,                         /* Report errors in .js files. */
     // "jsx": "react",                           /* Specify JSX code generation: 'preserve', 'react-native', or 'react'. */
-    "declaration": true /* Generates corresponding '.d.ts' file. */,
+    // "declaration": true /* Generates corresponding '.d.ts' file. */,
     // "sourceMap": true,                        /* Generates corresponding '.map' file. */
     // "outFile": "./",                       /* Concatenate and emit output to single file. */
-    "outDir": "./dist" /* Redirect output structure to the directory. */,
+    // "outDir": "./dist" /* Redirect output structure to the directory. */,
     // "rootDir": "./",                       /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */
     // "removeComments": true,                /* Do not emit comments to output. */
-    // "noEmit": true,                        /* Do not emit outputs. */
+    "noEmit": true /* Do not emit outputs. */,
     // "importHelpers": true,                 /* Import emit helpers from 'tslib'. */
     // "downlevelIteration": true,            /* Provide full support for iterables in 'for-of', spread, and destructuring when targeting 'ES5' or 'ES3'. */
     // "isolatedModules": true,               /* Transpile each file as a separate module (similar to 'ts.transpileModule'). */
@@ -40,9 +43,7 @@
 
     /* Module Resolution Options */
     "moduleResolution": "node" /* Specify module resolution strategy: 'node' (Node.js) or 'classic' (TypeScript pre-1.6). */,
-    "paths": {
-      "base64url": ["./node_modules/base64url/dist/base64url.d.ts"]
-    } /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */,
+    "paths": {} /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */,
     // "rootDirs": [],                        /* List of root folders whose combined content represents the structure of the project at runtime. */
     // "typeRoots": [],                       /* List of folders to include type definitions from. */
     // "types": [],                           /* Type declaration files to be included in compilation. */
@@ -58,14 +59,6 @@
     "experimentalDecorators": true /* Enables experimental support for ES7 decorators. */
     // "emitDecoratorMetadata": true,         /* Enables experimental support for emitting type metadata for decorators. */
   },
-  "include": ["./*.d.ts", "./**/*.ts"],
-  "exclude": [
-    "./artifacts",
-    "./node_modules",
-    "./dist/**/*",
-    "./scripts/*",
-    "./documentation/snippets",
-    "./documentation/examples",
-    "./documentation/scripts"
-  ]
+  "include": ["./documentation/scripts/*.ts", "./testing/helpers.ts", "./*.d.ts"],
+  "exclude": ["./artifacts", "./node_modules", "./dist/**/*", "./scripts/*"]
 }
